[fix] solution that actually works
But docs are out of date now. Not sure if I like it this way.
The advantage is that I don't have to parse the files for export
declarations. But the downside is it only works with default
exports (which we use but that's specific) and I guess the string
replace fuckery is pretty brittle